+++ flutter/lib/src/web/exceptions.dart
@@ -1,6 +1,5 @@
-// ignore_for_file: avoid_web_libraries_in_flutter
-
-import 'dart:js_util';
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import '../interface/exceptions.dart';
 import 'jason_wasm.dart' as wasm;
@@ -9,12 +8,21 @@ import 'jason_wasm.dart' as wasm;
 ///
 /// Returns `null` in case if the provided exception is not from Jason.
 String? _getName(dynamic e) {
-  try {
-    var exceptionConstructor = getProperty(e, 'constructor');
-    return getProperty(exceptionConstructor, 'name');
-  } catch (e) {
+  if (e is! JSObject) {
     return null;
   }
+
+  final constructor = e.getProperty('constructor'.toJS);
+  if (constructor is JSObject) {
+    if (constructor.hasProperty('name'.toJS).toDart) {
+      final name = constructor.getProperty('name'.toJS);
+      if (name is JSString) {
+        return name.toDart;
+      }
+    }
+  }
+
+  return null;
 }
 
 /// Converts the provided [wasm] exception into the Dart exception.