+++ flutter/lib/src/web/jason_wasm.dart
@@ -3,13 +3,12 @@
 @JS()
 library medea_jason;
 
-import 'dart:html' as html;
+import 'dart:js_interop';
 
-import 'package:js/js.dart';
-import 'package:js/js_util.dart' show promiseToFuture;
+import 'package:web/web.dart' as web;
 
 @JS()
-class FacingMode {
+extension type FacingMode._(JSObject _) implements JSObject {
   external static num get User;
   external static num get Environment;
   external static num get Left;
@@ -17,33 +16,33 @@ class FacingMode {
 }
 
 @JS()
-class MediaSourceKind {
+extension type MediaSourceKind._(JSObject _) implements JSObject {
   external static num get Device;
   external static num get Display;
 }
 
 @JS()
-class MediaKind {
+extension type MediaKind._(JSObject _) implements JSObject {
   external static num get Audio;
   external static num get Video;
 }
 
 @JS()
-class LocalMediaInitExceptionKind {
+extension type LocalMediaInitExceptionKind._(JSObject _) implements JSObject {
   external static num get GetUserMediaFailed;
   external static num get GetDisplayMediaFailed;
   external static num get LocalTrackIsEnded;
 }
 
 @JS()
-class RpcClientExceptionKind {
+extension type RpcClientExceptionKind._(JSObject _) implements JSObject {
   external static num get ConnectionLost;
   external static num get AuthorizationFailed;
   external static num get SessionFinished;
 }
 
 @JS()
-class AudioTrackConstraints {
+extension type AudioTrackConstraints._(JSObject _) implements JSObject {
   external void free();
   external factory AudioTrackConstraints();
   external void device_id(String device_id);
@@ -52,46 +51,20 @@ class AudioTrackConstraints {
 }
 
 @JS()
-class ConnectionHandle {
+extension type ConnectionHandle._(JSObject _) implements JSObject {
   external void free();
-  external void on_close(Function cb);
+  external void on_close(JSFunction cb);
   external String get_remote_member_id();
-  external void on_remote_track_added(Function cb);
-  external void on_quality_score_update(Function cb);
-}
-
-@JS('ConnectionHandle')
-abstract class _ConnectionHandle {
-  external Promise<dynamic> disable_remote_audio();
-  external Promise<dynamic> disable_remote_video(num? source_kind);
-  external Promise<dynamic> enable_remote_audio();
-  external Promise<dynamic> enable_remote_video(num? source_kind);
-}
-
-extension ConnectionHandleExtensions on ConnectionHandle {
-  Future<dynamic> disable_remote_audio() {
-    final tt = this as _ConnectionHandle;
-    return promiseToFuture(tt.disable_remote_audio());
-  }
-
-  Future<dynamic> disable_remote_video(num? source_kind) {
-    final tt = this as _ConnectionHandle;
-    return promiseToFuture(tt.disable_remote_video(source_kind));
-  }
-
-  Future<dynamic> enable_remote_audio() {
-    final tt = this as _ConnectionHandle;
-    return promiseToFuture(tt.enable_remote_audio());
-  }
-
-  Future<dynamic> enable_remote_video(num? source_kind) {
-    final tt = this as _ConnectionHandle;
-    return promiseToFuture(tt.enable_remote_video(source_kind));
-  }
+  external void on_remote_track_added(JSFunction cb);
+  external void on_quality_score_update(JSFunction cb);
+  external JSPromise<JSAny?> disable_remote_audio();
+  external JSPromise<JSAny?> disable_remote_video(num? source_kind);
+  external JSPromise<JSAny?> enable_remote_audio();
+  external JSPromise<JSAny?> enable_remote_video(num? source_kind);
 }
 
 @JS()
-class DeviceVideoTrackConstraints {
+extension type DeviceVideoTrackConstraints._(JSObject _) implements JSObject {
   external void free();
   external factory DeviceVideoTrackConstraints();
   external void device_id(String device_id);
@@ -108,7 +81,7 @@ class DeviceVideoTrackConstraints {
 }
 
 @JS()
-class DisplayVideoTrackConstraints {
+extension type DisplayVideoTrackConstraints._(JSObject _) implements JSObject {
   external void free();
   external factory DisplayVideoTrackConstraints();
   external void exact_height(num height);
@@ -120,20 +93,20 @@ class DisplayVideoTrackConstraints {
 }
 
 @JS()
-class EnumerateDevicesException {
+extension type EnumerateDevicesException._(JSObject _) implements JSObject {
   external void free();
-  external Error cause();
+  external JSObject cause();
   external String trace();
 }
 
 @JS()
-class FormatException {
+extension type FormatException._(JSObject _) implements JSObject {
   external void free();
   external String message();
 }
 
 @JS()
-class MediaDeviceDetails {
+extension type MediaDeviceDetails._(JSObject _) implements JSObject {
   external void free();
   external String device_id();
   external num kind();
@@ -142,15 +115,15 @@ class MediaDeviceDetails {
 }
 
 @JS()
-class InternalException {
+extension type InternalException._(JSObject _) implements JSObject {
   external void free();
   external String message();
-  external dynamic cause();
+  external String cause();
   external String trace();
 }
 
 @JS()
-class Jason {
+extension type Jason._(JSObject _) implements JSObject {
   external void free();
   external factory Jason();
   external RoomHandle init_room();
@@ -160,69 +133,44 @@ class Jason {
 }
 
 @JS()
-class LocalMediaInitException {
+extension type LocalMediaInitException._(JSObject _) implements JSObject {
   external void free();
   external num kind();
   external String message();
-  external dynamic cause();
+  external JSObject cause();
   external String trace();
 }
 
 @JS()
-class LocalMediaTrack {
+extension type LocalMediaTrack._(JSObject _) implements JSObject {
   external void free();
-  external html.MediaStreamTrack get_track();
+  external web.MediaStreamTrack get_track();
   external num kind();
   external num media_source_kind();
-  external void on_enabled(Function cb);
-}
-
-@JS('LocalMediaTrack')
-abstract class _LocalMediaTrack {
-  external Promise<dynamic> state();
-}
-
-extension LocalMediaTrackExtensions on LocalMediaTrack {
-  Future<dynamic> state() {
-    final tt = this as _LocalMediaTrack;
-    return promiseToFuture(tt.state());
-  }
+  external void on_enabled(JSFunction cb);
+  external JSPromise<JSAny?> state();
 }
 
 @JS()
-class MediaManagerHandle {
-  external void on_device_change(Function cb);
+extension type MediaManagerHandle._(JSObject _) implements JSObject {
+  external void on_device_change(JSFunction cb);
   external void free();
-}
-
-@JS('MediaManagerHandle')
-abstract class _MediaManagerHandle {
-  external Promise<List<dynamic>> enumerate_devices();
-  external Promise<List<dynamic>> init_local_tracks(MediaStreamSettings caps);
-}
-
-extension MediaManagerHandleExtensions on MediaManagerHandle {
-  Future<List<dynamic>> enumerate_devices() {
-    final tt = this as _MediaManagerHandle;
-    return promiseToFuture(tt.enumerate_devices());
-  }
-
-  Future<List<dynamic>> init_local_tracks(MediaStreamSettings caps) {
-    final tt = this as _MediaManagerHandle;
-    return promiseToFuture(tt.init_local_tracks(caps));
-  }
+  external JSPromise<JSArray<JSAny>> enumerate_devices();
+  external JSPromise<JSArray<JSAny>> init_local_tracks(
+    MediaStreamSettings caps,
+  );
 }
 
 @JS()
-class MediaSettingsUpdateException {
+extension type MediaSettingsUpdateException._(JSObject _) implements JSObject {
   external void free();
   external String message();
-  external dynamic cause();
+  external JSObject cause();
   external bool rolled_back();
 }
 
 @JS()
-class MediaStateTransitionException {
+extension type MediaStateTransitionException._(JSObject _) implements JSObject {
   external void free();
   external String message();
   external String trace();
@@ -230,7 +178,7 @@ class MediaStateTransitionException {
 }
 
 @JS()
-class MediaStreamSettings {
+extension type MediaStreamSettings._(JSObject _) implements JSObject {
   external void free();
   external factory MediaStreamSettings();
   external void audio(AudioTrackConstraints constraints);
@@ -238,51 +186,37 @@ class MediaStreamSettings {
   external void display_video(DisplayVideoTrackConstraints constraints);
 }
 
-@JS()
-class ReconnectHandle {
-  external void free();
-}
-
 @JS('ReconnectHandle')
-abstract class _ReconnectHandle {
-  external Promise<dynamic> reconnect_with_delay(num delay_ms);
-  external Promise<dynamic> reconnect_with_backoff(num starting_delay_ms,
-      num multiplier, num max_delay, num? max_elapsed_time_ms);
-}
-
-extension ReconnectHandleExtensions on ReconnectHandle {
-  Future<dynamic> reconnect_with_delay(num delay_ms) {
-    final tt = this as _ReconnectHandle;
-    return promiseToFuture(tt.reconnect_with_delay(delay_ms));
-  }
-
-  Future<dynamic> reconnect_with_backoff(num starting_delay_ms, num multiplier,
-      num max_delay, num? max_elapsed_time_ms) {
-    final tt = this as _ReconnectHandle;
-    return promiseToFuture(tt.reconnect_with_backoff(
-        starting_delay_ms, multiplier, max_delay, max_elapsed_time_ms));
-  }
+extension type ReconnectHandle._(JSObject _) implements JSObject {
+  external void free();
+  external JSPromise<JSAny?> reconnect_with_delay(num delay_ms);
+  external JSPromise<JSAny?> reconnect_with_backoff(
+    num starting_delay_ms,
+    num multiplier,
+    num max_delay,
+    num? max_elapsed_time_ms,
+  );
 }
 
 @JS()
-class RemoteMediaTrack {
+extension type RemoteMediaTrack._(JSObject _) implements JSObject {
   external void free();
-  external html.MediaStreamTrack get_track();
+  external web.MediaStreamTrack get_track();
   external bool enabled();
   external bool muted();
-  external void on_enabled(Function cb);
-  external void on_disabled(Function cb);
-  external void on_muted(Function cb);
-  external void on_unmuted(Function cb);
-  external void on_stopped(Function cb);
-  external void on_media_direction_changed(Function cb);
+  external void on_enabled(JSFunction cb);
+  external void on_disabled(JSFunction cb);
+  external void on_muted(JSFunction cb);
+  external void on_unmuted(JSFunction cb);
+  external void on_stopped(JSFunction cb);
+  external void on_media_direction_changed(JSFunction cb);
   external num kind();
   external num media_source_kind();
   external num media_direction();
 }
 
 @JS()
-class RoomCloseReason {
+extension type RoomCloseReason._(JSObject _) implements JSObject {
   external void free();
   external String reason();
   external bool is_closed_by_server();
@@ -290,128 +224,45 @@ class RoomCloseReason {
 }
 
 @JS()
-class RoomHandle {
+extension type RoomHandle._(JSObject _) implements JSObject {
   external void free();
-  external void on_new_connection(Function cb);
-  external void on_close(Function cb);
-  external void on_local_track(Function cb);
-  external void on_failed_local_media(Function cb);
-  external void on_connection_loss(Function cb);
-}
-
-@JS('RoomHandle')
-abstract class _RoomHandle {
-  external Promise<dynamic> join(String token);
-  external Promise<dynamic> set_local_media_settings(
-      MediaStreamSettings settings, bool stop_first, bool rollback_on_fail);
-  external Promise<dynamic> mute_audio();
-  external Promise<dynamic> unmute_audio();
-  external Promise<dynamic> mute_video(num? source_kind);
-  external Promise<dynamic> unmute_video(num? source_kind);
-  external Promise<dynamic> disable_audio();
-  external Promise<dynamic> enable_audio();
-  external Promise<dynamic> disable_video(num? source_kind);
-  external Promise<dynamic> enable_video(num? source_kind);
-  external Promise<dynamic> disable_remote_audio();
-  external Promise<dynamic> disable_remote_video(num? source_kind);
-  external Promise<dynamic> enable_remote_audio();
-  external Promise<dynamic> enable_remote_video(num? source_kind);
-}
-
-extension RoomHandleExtensions on RoomHandle {
-  Future<dynamic> join(String token) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.join(token));
-  }
-
-  Future<dynamic> set_local_media_settings(
-      MediaStreamSettings settings, bool stop_first, bool rollback_on_fail) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(
-        tt.set_local_media_settings(settings, stop_first, rollback_on_fail));
-  }
-
-  Future<dynamic> mute_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.mute_audio());
-  }
-
-  Future<dynamic> unmute_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.unmute_audio());
-  }
-
-  Future<dynamic> mute_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.mute_video(source_kind));
-  }
-
-  Future<dynamic> unmute_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.unmute_video(source_kind));
-  }
-
-  Future<dynamic> disable_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.disable_audio());
-  }
-
-  Future<dynamic> enable_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.enable_audio());
-  }
-
-  Future<dynamic> disable_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.disable_video(source_kind));
-  }
-
-  Future<dynamic> enable_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.enable_video(source_kind));
-  }
-
-  Future<dynamic> disable_remote_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.disable_remote_audio());
-  }
-
-  Future<dynamic> disable_remote_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.disable_remote_video(source_kind));
-  }
-
-  Future<dynamic> enable_remote_audio() {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.enable_remote_audio());
-  }
-
-  Future<dynamic> enable_remote_video(num? source_kind) {
-    final tt = this as _RoomHandle;
-    return promiseToFuture(tt.enable_remote_video(source_kind));
-  }
+  external void on_new_connection(JSFunction cb);
+  external void on_close(JSFunction cb);
+  external void on_local_track(JSFunction cb);
+  external void on_failed_local_media(JSFunction cb);
+  external void on_connection_loss(JSFunction cb);
+  external JSPromise<JSAny?> join(String token);
+  external JSPromise<JSAny?> set_local_media_settings(
+    MediaStreamSettings settings,
+    bool stop_first,
+    bool rollback_on_fail,
+  );
+  external JSPromise<JSAny?> mute_audio();
+  external JSPromise<JSAny?> unmute_audio();
+  external JSPromise<JSAny?> mute_video(num? source_kind);
+  external JSPromise<JSAny?> unmute_video(num? source_kind);
+  external JSPromise<JSAny?> disable_audio();
+  external JSPromise<JSAny?> enable_audio();
+  external JSPromise<JSAny?> disable_video(num? source_kind);
+  external JSPromise<JSAny?> enable_video(num? source_kind);
+  external JSPromise<JSAny?> disable_remote_audio();
+  external JSPromise<JSAny?> disable_remote_video(num? source_kind);
+  external JSPromise<JSAny?> enable_remote_audio();
+  external JSPromise<JSAny?> enable_remote_video(num? source_kind);
 }
 
 @JS()
-class RpcClientException {
+extension type RpcClientException._(JSObject _) implements JSObject {
   external void free();
   external num kind();
   external String message();
-  external dynamic cause();
+  external JSObject cause();
   external String trace();
 }
 
 @JS()
-class StateError {
+extension type StateError._(JSObject _) implements JSObject {
   external void free();
   external String message();
   external String trace();
-}
-
-@JS()
-abstract class Promise<T> {
-  external factory Promise(
-      void Function(void Function(T result) resolve, Function reject) executor);
-  external Promise then(void Function(T result) onFulfilled,
-      [Function onRejected]);
 }